requirements.txt
setuptools==38.4.0
PyJWT==1.6.1
pyzmq==17.0.0
